+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 1.18.0
 * `NEW` `LuaDoc`: supports `---@diagnostic disable`
+* `NEW` code-action: convert JSON to Lua
 * `CHG` `Windows`: dose not provide `ucrt` any more
 * `CHG` `Lua.workspace.library`: use `path[]` instead of `<path, true>`
 * `FIX` missed syntax error `local a <const>= 1`